refactor(CANDriver): use uint32_t for rxId/txId with UINT32_MAX sentinel
Replace int32_t rxId/txId (using -1 as sentinel) with uint32_t and a
named noId = UINT32_MAX constant. UINT32_MAX is never a valid CAN ID
(max 29-bit = 0x1FFFFFFF), so it is a clean sentinel. Removes all
static_cast<int32_t> and static_cast<uint32_t> casts around these fields
and aligns packetId() return type with its callers (uint32_t extId).

+++ lib/CANDriver/src/ESP32SJA1000.cpp
@@ -151,7 +151,7 @@ uint8_t ESP32SJA1000::endPacket() {
 
   uint8_t dataReg;
 
-  const uint32_t id = static_cast<uint32_t>(txId);
+  const uint32_t id = txId;
   if(txExtended) {
     writeRegister(regEff, static_cast<uint8_t>(0x80U | (txRtr ? 0x40U : 0x00U) | (0x0FU & txLength)));
     writeRegister(static_cast<uint8_t>(regEff + 1U), static_cast<uint8_t>(id >> 21));
@@ -202,17 +202,17 @@ uint8_t ESP32SJA1000::parsePacket() {
   uint8_t dataReg;
 
   if(rxExtended) {
-    rxId = static_cast<int32_t>(
+    rxId =
       (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regEff + 1U))) << 21U) |
       (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regEff + 2U))) << 13U) |
       (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regEff + 3U))) << 5U)  |
-      (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regEff + 4U))) >> 3U));
+      (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regEff + 4U))) >> 3U);
 
     dataReg = static_cast<uint8_t>(regEff + 5U);
   } else {
-    rxId = static_cast<int32_t>(
+    rxId =
       (static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regSff + 1U))) << 3U) |
-      ((static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regSff + 2U))) >> 5U) & 0x07U));
+      ((static_cast<uint32_t>(readRegister(static_cast<uint8_t>(regSff + 2U))) >> 5U) & 0x07U);
 
     dataReg = static_cast<uint8_t>(regSff + 3U);
   }
